Add dependency checking to scala_compile.
Adds an optional analysis pass post-compilation for scala
targets that uses real class dependencies to generate a list
of the scala targets that any compilation target actually depends
on, and then compares that to the declared list of dependencies
to detect any spurious dependencies that are accidentally permitted
by the chunked compiler.

The new analysis is only run if the flag "check_missing_deps" is
passed to the scala compiler. (e.g., by running "pants compile --compile-scalac-check_missing_deps").